Move the allocator related functions to allocator.c
